+++ src/LayeredCanvas.h
@@ -28,9 +28,9 @@ namespace hg {
     class LayeredCanvas {
     public:
         LayeredCanvas(Canvas& canvas);
-        void drawRect(int layer, int x, int y, int width, int height, unsigned colour);
-        void drawLine(int layer, int xa, int ya, int xb, int yb, int width, unsigned colour);
-        void drawText(int layer, std::string const& text, int x, int y, int size, unsigned colour);
+        void drawRect(int layer, float x, float y, float width, float height, unsigned colour);
+        void drawLine(int layer, float xa, float ya, float xb, float yb, float width, unsigned colour);
+        void drawText(int layer, std::string const& text, float x, float y, float size, unsigned colour);
         Flusher getFlusher();
         
         ~LayeredCanvas();
